LPS-125095 Move classes that implement FormNavigatorEntry to package *.frontend.taglib.form.navigator
