+++ modules/apps/commerce/commerce-product-content-web/src/main/java/com/liferay/commerce/product/content/web/internal/product/compare/mini/frontend/taglib/form/navigator/ADTFormNavigatorEntry.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.commerce.product.content.web.internal.product.compare.servlet.taglib.ui;
+package com.liferay.commerce.product.content.web.internal.product.compare.mini.frontend.taglib.form.navigator;
 
-import com.liferay.commerce.product.content.web.internal.configuration.CPCompareContentPortletInstanceConfiguration;
-import com.liferay.commerce.product.content.web.internal.constants.CPCompareContentConstants;
+import com.liferay.commerce.product.content.web.internal.configuration.CPCompareContentMiniPortletInstanceConfiguration;
+import com.liferay.commerce.product.content.web.internal.constants.CPCompareContentMiniConstants;
 import com.liferay.frontend.taglib.form.navigator.BaseJSPFormNavigatorEntry;
 import com.liferay.frontend.taglib.form.navigator.FormNavigatorEntry;
 import com.liferay.portal.kernel.exception.PortalException;
@@ -46,12 +46,12 @@ public class ADTFormNavigatorEntry extends BaseJSPFormNavigatorEntry<Void> {
 
 	@Override
 	public String getCategoryKey() {
-		return CPCompareContentConstants.CATEGORY_KEY_RENDER_SELECTION;
+		return CPCompareContentMiniConstants.CATEGORY_KEY_RENDER_SELECTION;
 	}
 
 	@Override
 	public String getFormNavigatorId() {
-		return CPCompareContentConstants.FORM_NAVIGATOR_ID_CONFIGURATION;
+		return CPCompareContentMiniConstants.FORM_NAVIGATOR_ID_CONFIGURATION;
 	}
 
 	@Override
@@ -80,7 +80,7 @@ public class ADTFormNavigatorEntry extends BaseJSPFormNavigatorEntry<Void> {
 
 	@Override
 	protected String getJspPath() {
-		return "/compare_products/configuration/adt.jsp";
+		return "/compare_products_mini/configuration/adt.jsp";
 	}
 
 	private boolean _isSelectionStyleADT() {
@@ -92,13 +92,14 @@ public class ADTFormNavigatorEntry extends BaseJSPFormNavigatorEntry<Void> {
 		PortletDisplay portletDisplay = themeDisplay.getPortletDisplay();
 
 		try {
-			CPCompareContentPortletInstanceConfiguration
-				cpCompareContentPortletInstanceConfiguration =
+			CPCompareContentMiniPortletInstanceConfiguration
+				cpCompareContentMiniPortletInstanceConfiguration =
 					portletDisplay.getPortletInstanceConfiguration(
-						CPCompareContentPortletInstanceConfiguration.class);
+						CPCompareContentMiniPortletInstanceConfiguration.class);
 
 			String selectionStyle =
-				cpCompareContentPortletInstanceConfiguration.selectionStyle();
+				cpCompareContentMiniPortletInstanceConfiguration.
+					selectionStyle();
 
 			return selectionStyle.equals(getKey());
 		}
