+++ modules/apps/commerce/commerce-product-content-web/src/main/java/com/liferay/commerce/product/content/web/internal/product/compare/mini/frontend/taglib/form/navigator/ProductListRendererFormNavigatorEntry.java
@@ -12,22 +12,25 @@
  * details.
  */
 
-package com.liferay.commerce.product.content.web.internal.product.publisher.servlet.taglib.ui;
+package com.liferay.commerce.product.content.web.internal.product.compare.mini.frontend.taglib.form.navigator;
 
-import com.liferay.commerce.product.content.web.internal.constants.CPPublisherConstants;
+import com.liferay.commerce.product.content.web.internal.configuration.CPCompareContentMiniPortletInstanceConfiguration;
+import com.liferay.commerce.product.content.web.internal.constants.CPCompareContentMiniConstants;
 import com.liferay.frontend.taglib.form.navigator.BaseJSPFormNavigatorEntry;
 import com.liferay.frontend.taglib.form.navigator.FormNavigatorEntry;
+import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
 import com.liferay.portal.kernel.theme.PortletDisplay;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
 import java.util.Locale;
-
-import javax.portlet.PortletPreferences;
+import java.util.ResourceBundle;
 
 import javax.servlet.ServletContext;
 
@@ -38,34 +41,38 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alessio Antonio Rendina
  */
 @Component(
-	enabled = false, property = "form.navigator.entry.order:Integer=650",
+	enabled = false, property = "form.navigator.entry.order:Integer=500",
 	service = FormNavigatorEntry.class
 )
-public class ADTFormNavigatorEntry extends BaseJSPFormNavigatorEntry<Void> {
+public class ProductListRendererFormNavigatorEntry
+	extends BaseJSPFormNavigatorEntry<Void> {
 
 	@Override
 	public String getCategoryKey() {
-		return CPPublisherConstants.CATEGORY_KEY_RENDER_SELECTION;
+		return CPCompareContentMiniConstants.CATEGORY_KEY_RENDER_SELECTION;
 	}
 
 	@Override
 	public String getFormNavigatorId() {
-		return CPPublisherConstants.FORM_NAVIGATOR_ID_CONFIGURATION;
+		return CPCompareContentMiniConstants.FORM_NAVIGATOR_ID_CONFIGURATION;
 	}
 
 	@Override
 	public String getKey() {
-		return "adt";
+		return "product-list-renderer";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "display-template");
+		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
+			"content.Language", locale, getClass());
+
+		return LanguageUtil.get(resourceBundle, getKey());
 	}
 
 	@Override
 	public boolean isVisible(User user, Void object) {
-		return _isSelectionStyleADT();
+		return _isSelectionStyleCustomRenderer();
 	}
 
 	@Override
@@ -79,10 +86,10 @@ public class ADTFormNavigatorEntry extends BaseJSPFormNavigatorEntry<Void> {
 
 	@Override
 	protected String getJspPath() {
-		return "/product_publisher/configuration/adt.jsp";
+		return "/compare_products_mini/configuration/product_list_renderer.jsp";
 	}
 
-	private boolean _isSelectionStyleADT() {
+	private boolean _isSelectionStyleCustomRenderer() {
 		ServiceContext serviceContext =
 			ServiceContextThreadLocal.getServiceContext();
 
@@ -90,18 +97,28 @@ public class ADTFormNavigatorEntry extends BaseJSPFormNavigatorEntry<Void> {
 
 		PortletDisplay portletDisplay = themeDisplay.getPortletDisplay();
 
-		PortletPreferences portletPreferences =
-			themeDisplay.getStrictLayoutPortletSetup(
-				themeDisplay.getLayout(), portletDisplay.getPortletResource());
+		try {
+			CPCompareContentMiniPortletInstanceConfiguration
+				cpCompareContentMiniPortletInstanceConfiguration =
+					portletDisplay.getPortletInstanceConfiguration(
+						CPCompareContentMiniPortletInstanceConfiguration.class);
 
-		String renderSelection = GetterUtil.getString(
-			portletPreferences.getValue("renderSelection", null), "custom");
+			String selectionStyle =
+				cpCompareContentMiniPortletInstanceConfiguration.
+					selectionStyle();
 
-		if (renderSelection.equals("adt")) {
-			return true;
+			return selectionStyle.equals("custom");
 		}
+		catch (PortalException portalException) {
+			if (_log.isDebugEnabled()) {
+				_log.debug(portalException, portalException);
+			}
 
-		return false;
+			return false;
+		}
 	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		ProductListRendererFormNavigatorEntry.class);
 
 }
